Add global feedback widget to all AI help buttons
- Create shared AIHelpFeedbackPanel component for collecting feedback
- Update AIHelpButton and AIHelpIconButton to use shared panel
- Update AIHelpSubmissionErrorButton to show feedback after copying
- Update AIHelpTestErrorButton to show feedback after copying
- Add new migration to support additional context types:
  test_failure, build_error, test_insights
- Update RPC to validate submissions for test_failure/build_error
