+++ components/ai-help/AIHelpButton.tsx
@@ -4,12 +4,10 @@ import { Button } from "@/components/ui/button";
 import { Tooltip } from "@/components/ui/tooltip";
 import { toaster } from "@/components/ui/toaster";
 import { useIsGraderOrInstructor } from "@/hooks/useClassProfiles";
-import { aiHelpFeedbackSubmit } from "@/lib/edgeFunctions";
-import { createClient } from "@/utils/supabase/client";
-import { Box, HStack, Icon, IconButton, Input, Text, Textarea, VStack } from "@chakra-ui/react";
+import { Box, HStack, Icon, IconButton, Input, Text } from "@chakra-ui/react";
 import { useCallback, useMemo, useState } from "react";
 import { BsRobot, BsCopy, BsX } from "react-icons/bs";
-import { LuThumbsUp, LuThumbsDown } from "react-icons/lu";
+import { AIHelpFeedbackPanel } from "./AIHelpFeedbackPanel";
 
 /**
  * Props for the AIHelpButton component
@@ -145,151 +143,6 @@ Provide helpful guidance that:
 }
 
 /**
- * Submit feedback via Edge Function
- */
-async function submitFeedback(
-  classId: number,
-  contextType: "help_request" | "discussion_thread",
-  resourceId: number,
-  rating: "thumbs_up" | "thumbs_down",
-  comment?: string
-): Promise<boolean> {
-  try {
-    const supabase = createClient();
-    await aiHelpFeedbackSubmit(
-      {
-        class_id: classId,
-        context_type: contextType,
-        resource_id: resourceId,
-        rating,
-        comment: comment || undefined
-      },
-      supabase
-    );
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-/**
- * Feedback component shown after copying AI context
- */
-function FeedbackPanel({
-  classId,
-  contextType,
-  resourceId,
-  onClose
-}: {
-  classId: number;
-  contextType: "help_request" | "discussion_thread";
-  resourceId: number;
-  onClose: () => void;
-}) {
-  const [rating, setRating] = useState<"thumbs_up" | "thumbs_down" | null>(null);
-  const [comment, setComment] = useState("");
-  const [submitting, setSubmitting] = useState(false);
-  const [submitted, setSubmitted] = useState(false);
-
-  const handleSubmit = async () => {
-    if (!rating) return;
-
-    setSubmitting(true);
-    const success = await submitFeedback(classId, contextType, resourceId, rating, comment);
-    setSubmitting(false);
-
-    if (success) {
-      setSubmitted(true);
-      toaster.success({
-        title: "Thanks for your feedback!",
-        description: "Your feedback helps us improve the AI assistance feature."
-      });
-    } else {
-      toaster.error({
-        title: "Failed to submit feedback",
-        description: "Please try again later."
-      });
-    }
-  };
-
-  if (submitted) {
-    return (
-      <Box p={3} borderWidth="1px" borderRadius="md" bg="green.subtle" maxW="400px">
-        <HStack justify="space-between">
-          <Text fontSize="sm" fontWeight="medium" color="green.fg">
-            Thank you for your feedback!
-          </Text>
-          <IconButton aria-label="Close" size="xs" variant="ghost" onClick={onClose}>
-            <Icon as={BsX} />
-          </IconButton>
-        </HStack>
-      </Box>
-    );
-  }
-
-  return (
-    <Box p={3} borderWidth="1px" borderRadius="md" bg="bg.subtle" maxW="400px">
-      <HStack justify="space-between" mb={2}>
-        <Text fontSize="sm" fontWeight="medium">
-          How was the AI assistance?
-        </Text>
-        <IconButton aria-label="Close" size="xs" variant="ghost" onClick={onClose}>
-          <Icon as={BsX} />
-        </IconButton>
-      </HStack>
-
-      <VStack gap={3} align="stretch">
-        <HStack gap={2} justify="center">
-          <IconButton
-            aria-label="Thumbs up"
-            size="lg"
-            variant={rating === "thumbs_up" ? "solid" : "outline"}
-            colorPalette={rating === "thumbs_up" ? "green" : "gray"}
-            onClick={() => setRating("thumbs_up")}
-          >
-            <Icon as={LuThumbsUp} boxSize={5} />
-          </IconButton>
-          <IconButton
-            aria-label="Thumbs down"
-            size="lg"
-            variant={rating === "thumbs_down" ? "solid" : "outline"}
-            colorPalette={rating === "thumbs_down" ? "red" : "gray"}
-            onClick={() => setRating("thumbs_down")}
-          >
-            <Icon as={LuThumbsDown} boxSize={5} />
-          </IconButton>
-        </HStack>
-
-        <Textarea
-          placeholder="Any additional feedback? (optional)"
-          value={comment}
-          onChange={(e) => setComment(e.target.value)}
-          fontSize="sm"
-          rows={2}
-          maxLength={2000}
-        />
-
-        <HStack gap={2}>
-          <Button size="sm" variant="ghost" onClick={onClose} flex={1}>
-            Skip
-          </Button>
-          <Button
-            size="sm"
-            colorPalette="purple"
-            onClick={handleSubmit}
-            disabled={!rating || submitting}
-            loading={submitting}
-            flex={1}
-          >
-            Submit
-          </Button>
-        </HStack>
-      </VStack>
-    </Box>
-  );
-}
-
-/**
  * AIHelpButton component for launching AI assistance context
  *
  * This component provides a button that instructors and graders can use
@@ -350,7 +203,9 @@ export function AIHelpButton({
   }
 
   if (showFeedback) {
-    return <FeedbackPanel classId={classId} contextType={contextType} resourceId={resourceId} onClose={handleClose} />;
+    return (
+      <AIHelpFeedbackPanel classId={classId} contextType={contextType} resourceId={resourceId} onClose={handleClose} />
+    );
   }
 
   if (showContext) {
@@ -448,7 +303,7 @@ export function AIHelpIconButton({
 
   if (showFeedback) {
     return (
-      <FeedbackPanel
+      <AIHelpFeedbackPanel
         classId={classId}
         contextType={contextType}
         resourceId={resourceId}
